Split HttpConnectionPool into multiple files
